+++ tests/meen_hw_test/source/MeenHwGTest.cpp
@@ -20,9 +20,8 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
 
-#include <gtest/gtest.h>
-
 #include <bit>
+#include <gtest/gtest.h>
 #include <vector>
 
 #include "meen_hw/MH_Factory.h"
@@ -33,7 +32,7 @@ namespace meen_hw::tests
 	class MeenHwTest : public testing::Test
 	{
 	protected:
-		static std::unique_ptr<MH_II8080ArcadeIO> i8080ArcadeIO_; 
+		static std::unique_ptr<MH_II8080ArcadeIO> i8080ArcadeIO_;
 	public:
 		static void SetUpTestCase();
 	};
@@ -52,11 +51,11 @@ namespace meen_hw::tests
 
 	TEST_F(MeenHwTest, Version)
 	{
-		EXPECT_NE(nullptr, Version());	
+		EXPECT_NE(nullptr, Version());
 	}
 
 	TEST_F(MeenHwTest, ResourcePool)
-	{		
+	{
 		static int counter = 0;
 
 		struct ResourceDeleter
@@ -113,7 +112,7 @@ namespace meen_hw::tests
 			ASSERT_NE(nullptr, outlivePool);
 			*outlivePool = 42;
 		}
-		
+
 		// The pool is dead, resource should be valid
 		EXPECT_EQ(42, *outlivePool);
 		// The pool is dead, the resource should be destroyed
@@ -156,7 +155,7 @@ namespace meen_hw::tests
 			EXPECT_EQ(0x00, value);
 			// Try to play audio sound effects 2 and 3
 			value = i8080ArcadeIO_->WritePort(port, 0x06);
-			// We should get audio sound effect 3 but not 2 
+			// We should get audio sound effect 3 but not 2
 			EXPECT_EQ(0x04, value);
 			// Try to play audio sound effect 3 (note this does not include 2)
 			value = i8080ArcadeIO_->WritePort(port, 0x04);
@@ -164,7 +163,7 @@ namespace meen_hw::tests
 			EXPECT_EQ(0x00, value);
 			// Try to play audio sound effects 2 and 3
 			value = i8080ArcadeIO_->WritePort(port, 0x06);
-			// We should get the audio sound effect 2 but not 3 
+			// We should get the audio sound effect 2 but not 3
 			EXPECT_EQ(0x02, value);
 			// Don't play any audio sound effects
 			value = i8080ArcadeIO_->WritePort(port, 0x00);
@@ -244,11 +243,20 @@ namespace meen_hw::tests
 
 	TEST_F(MeenHwTest, SetOptions)
 	{
-		EXPECT_THROW(i8080ArcadeIO_->SetOptions("{\"bpp\":2}"), std::invalid_argument);
-		EXPECT_THROW(i8080ArcadeIO_->SetOptions("{\"colour\":\"black\" }"), std::invalid_argument);
-		EXPECT_THROW(i8080ArcadeIO_->SetOptions("{\"orientation\":\"up\"}"), std::invalid_argument);
-		EXPECT_THROW(i8080ArcadeIO_->SetOptions("{\"invalid-option\":1}"), std::invalid_argument);
-		EXPECT_NO_THROW(i8080ArcadeIO_->SetOptions("{\"bpp\":8,\"colour\":\"random\",\"orientation\":\"cocktail\"}"));
+		auto checkErrc = [](const std::error_code& ec, bool success, const char* expectedMsg)
+		{
+			if (success == true) EXPECT_FALSE(ec); else EXPECT_TRUE(ec);
+			EXPECT_EQ(expectedMsg, ec.message());
+		};
+
+		EXPECT_NO_THROW
+		(
+			checkErrc(i8080ArcadeIO_->SetOptions("{\"bpp\":2}"), false, "The bpp configuration option is invalid");
+			checkErrc(i8080ArcadeIO_->SetOptions("{\"colour\":\"black\" }"), false, "The colour configuration option is invalid");
+			checkErrc(i8080ArcadeIO_->SetOptions("{\"orientation\":\"up\"}"), false, "The orientation configuration parameter is invalid");
+			checkErrc(i8080ArcadeIO_->SetOptions("syntax-error"), false, "A json parse error occurred while processing the configuration file");
+			checkErrc(i8080ArcadeIO_->SetOptions("{\"bpp\":8,\"colour\":\"random\",\"orientation\":\"cocktail\"}"), true, "Success");
+		);
 	}
 
 	TEST_F(MeenHwTest, GetVRAMDimensions)
@@ -265,34 +273,7 @@ namespace meen_hw::tests
 	TEST_F(MeenHwTest, BlitVRAM)
 	{
 		uint8_t srcVRAM[7168]; // 7168 - width * height @ 1bpp
-		uint8_t expectedVRAMUpright1[7168]; // 57344 - width * height @ 1pp
-		uint8_t expectedVRAMCocktail8[57344]; // 57344 - width * height @ 8pp
-		uint8_t expectedVRAMUpright8[57344]; // 57344 - width * height @ 8pp
-		
-		// We need to output white (0xFF) in the uncompressed case
-		i8080ArcadeIO_->SetOptions("{\"colour\":\"white\"}");
-
-		// Set the src vram to be blitted to be an alternating black and white scanline pattern
-		// This will act as the expectedVRAM for 1bpp native orientation test
-		for (auto data = srcVRAM; data < srcVRAM + 7168; data += 64)
-		{
-			// 32 - compressed row bytes
-			std::fill_n(data, 32, 0x00);
-			std::fill_n(data + 32, 32, 0xFF);
-		}
-
-		// Vertical black and white bars
-		std::fill(expectedVRAMUpright1, expectedVRAMUpright1 + 7168, 0xAA);
-
-		for (auto data = expectedVRAMCocktail8; data < expectedVRAMCocktail8 + 57344; data += 512)
-		{
-			// 256 - uncompressed row bytes
-			std::fill_n(data, 256, 0x00);
-			std::fill_n(data + 256, 256, 0xFF);
-		}
-
-		auto data = expectedVRAMUpright8;
-		std::fill_n(std::bit_cast<uint16_t*>(data), 28672, 0xFF00);
+		uint8_t expectedVRAM[57344]; // 57344 - width * height @ 8pp
 
 		auto checkVRAM = [this](std::span<uint8_t> VRAMToBlit, std::span<uint8_t> expectedVRAM, int expectedRowBytes, int padding, int compressed, const char* options)
 		{
@@ -313,25 +294,50 @@ namespace meen_hw::tests
 			}
 		};
 
+		// We need to output white (0xFF) in the uncompressed case
+		EXPECT_FALSE(i8080ArcadeIO_->SetOptions("{\"colour\":\"white\"}"));
+
+		// Set the src vram to be blitted to be an alternating black and white scanline pattern
+		// This will act as the expectedVRAM for 1bpp native orientation test
+		for (auto data = srcVRAM; data < srcVRAM + 7168; data += 64)
+		{
+			// 32 - compressed row bytes
+			std::fill_n(data, 32, 0x00);
+			std::fill_n(data + 32, 32, 0xFF);
+		}
+
 		// Native blit without padding
 		checkVRAM(std::span(srcVRAM), std::span(srcVRAM), 32, 0, 3, "{\"bpp\":1,\"orientation\":\"cocktail\"}");
 		// Native blit with padding
 		checkVRAM(std::span(srcVRAM), std::span(srcVRAM), 32, 2, 3, "{\"bpp\":1,\"orientation\":\"cocktail\"}");
 
-		// Native orientation 8pp blit without padding
-		checkVRAM(std::span(srcVRAM), std::span(expectedVRAMCocktail8), 256, 0, 0, "{\"bpp\":8,\"orientation\":\"cocktail\"}");
-		// Native orientation 8pp blit with padding
-		checkVRAM(std::span(srcVRAM), std::span(expectedVRAMCocktail8), 256, 16, 0, "{\"bpp\":8,\"orientation\":\"cocktail\"}");
+		// Vertical black and white bars
+		std::fill(expectedVRAM, expectedVRAM + 7168, 0xAA);
 
 		// Native bpp blit with upright orientation without padding
-		checkVRAM(std::span(srcVRAM), std::span(expectedVRAMUpright1), 28, 0, 3, "{\"bpp\":1,\"orientation\":\"upright\"}");
+		checkVRAM(std::span(srcVRAM), std::span(expectedVRAM, 7168), 28, 0, 3, "{\"bpp\":1,\"orientation\":\"upright\"}");
 		// Native bpp blit with upright orientation with padding
-		checkVRAM(std::span(srcVRAM), std::span(expectedVRAMUpright1), 28, 2, 3, "{\"bpp\":1,\"orientation\":\"upright\"}");
+		checkVRAM(std::span(srcVRAM), std::span(expectedVRAM, 7168), 28, 2, 3, "{\"bpp\":1,\"orientation\":\"upright\"}");
+
+		for (auto data = expectedVRAM; data < expectedVRAM + 57344; data += 512)
+		{
+			// 256 - uncompressed row bytes
+			std::fill_n(data, 256, 0x00);
+			std::fill_n(data + 256, 256, 0xFF);
+		}
+
+		// Native orientation 8pp blit without padding
+		checkVRAM(std::span(srcVRAM), std::span(expectedVRAM), 256, 0, 0, "{\"bpp\":8,\"orientation\":\"cocktail\"}");
+		// Native orientation 8pp blit with padding
+		checkVRAM(std::span(srcVRAM), std::span(expectedVRAM), 256, 16, 0, "{\"bpp\":8,\"orientation\":\"cocktail\"}");
+
+		auto data = expectedVRAM;
+		std::fill_n(std::bit_cast<uint16_t*>(data), 28672, 0xFF00);
 
 		// 8 bpp blit with upright orientation without padding
-		checkVRAM(std::span(srcVRAM), std::span(expectedVRAMUpright8), 224, 0, 0, "{\"bpp\":8,\"orientation\":\"upright\"}");
+		checkVRAM(std::span(srcVRAM), std::span(expectedVRAM), 224, 0, 0, "{\"bpp\":8,\"orientation\":\"upright\"}");
 		// 8 bpp blit with upright orientation with padding
-		checkVRAM(std::span(srcVRAM), std::span(expectedVRAMUpright8), 224, 16, 0, "{\"bpp\":8,\"orientation\":\"upright\"}");
+		checkVRAM(std::span(srcVRAM), std::span(expectedVRAM), 224, 16, 0, "{\"bpp\":8,\"orientation\":\"upright\"}");
 	}
 #endif